DON-998: Refactor, merge MandateService with RegularGivingService
Didn't realise I had accidentally created two separate service classes
